+++ golang/testsuite/testsuite_impl/basic_datastore_and_api_test/basic_datastore_and_api_test_.go
@@ -180,10 +180,7 @@ func getApiServiceContainerConfigSupplier(datastoreClient *datastore_service_cli
 }
 
 func createDatastoreConfigFileInServiceDirectory(datastoreClient *datastore_service_client.DatastoreClient, sharedDirectory *services.SharedPath) (*services.SharedPath, error) {
-	configFileFilePath, err := sharedDirectory.GetChildPath(configFileKey)
-	if err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred getting file object '%v' from shared directory", configFileKey)
-	}
+	configFileFilePath := sharedDirectory.GetChildPath(configFileKey)
 
 	logrus.Infof("Config file absolute path on this container: %v , on service container: %v", configFileFilePath.GetAbsPathOnThisContainer(), configFileFilePath.GetAbsPathOnServiceContainer())