+++ docs/changelog.md
@@ -1,11 +1,14 @@
 # TBD
+### Features
+* Upgraded to [Kurtosis Testsuite API Lib 0.9.0](https://github.com/kurtosis-tech/kurtosis-testsuite-api-lib/blob/develop/docs/changelog.md#090)
+
 ### Fixes
 * Fixed an issue where the Docker image would get rebuilt even if just a Vim swapfile changed
 
 # 1.33.4
 ### Features
 * Upgraded to [Kurt Client API 0.17.0](https://github.com/kurtosis-tech/kurtosis-client/blob/develop/docs/changelog.md#0170)
-* Upgraded to [Kurtosis Testsuite API Lib 0.8.0](https://github.com/kurtosis-tech/kurtosis-testsuite-api-lib/blob/develop/docs/changelog.md#081)
+* Upgraded to [Kurtosis Testsuite API Lib 0.8.0](https://github.com/kurtosis-tech/kurtosis-testsuite-api-lib/blob/develop/docs/changelog.md#080)
 
 # 1.33.3
 ### Fixes